Variable label sanitized
